parse and apply "expand <increment>" to hulls independent of smoothing
This was previously coded as a sub-option to "smooth path":
	plot ... convex hull smooth path expand <scale-factor>

Now it is accepted anywhere after "convexhull" regardless of whether
"smooth path" is also chosen.

Furthermore, the <increment> is now interpreted as constant value
by which each vertex is displaced.  Previously it was interpreted
as a scale factor.

The "expand" option remains EXPERIMENTAL.
In particular the plan is to replace the expand_hull() algorithm
with a polygon offset algorithm that would work for both convex
and concave hulls.