plot "with labels" must not use data filters
Filter operations replace the array of original data points
plot->points with a new array of constructed or rearranged points.
However labels are kept in a separate structure that was ignored
(or truncated) by the filter, so they no longer correspond to the
points in plot->points.

We now give an error on any combination of filter + "with labels".
Possibly in the future it might be desirable to apply some particular
filter to labels, but it will have to somehow restructure the labels
to match the restructured data points.