reimplement hull expansion to use polygon offsetting
This replaces the current implementation of "expand <value>"
applied to hulls.  The current method pushes each hull point
directly away from the center of mass.
For convex hulls that is OK although not perfect.
For concave hulls it can be quite bad, even pushing points into
the interior of the cluster rather than out into the exterior.

This replacement instead shifts each edge of the hull away from
the local interior by displacing it along the edge normal.

This commit
- Removes the unused cluster outlier routine
  Removes code implementing the previous expansion method
- Consolidates determination of the min/max/centroid
  into a routine cluster_stats() used for both convex and concave
  hull determination and expansion
- Reworks expand_hull() to implement hull expansion as a
  polygon offset operation