Delaunay triangulation: start with a rectangle rather than a triangle
Triangulating data for which xrange and yrange were greatly different
revealed that the "sufficiently big" starting triangle was not, in fact,
sufficient big.  We try to address this by
(1) Making it actually a bounding box rather than a triangle
(2) Increasing the buffer to 100 * max(xrange,yrange)

I imagine there still can be pathological cases, but this fixes
the problem for test data with a 1000-fold difference in x and y
ranges.