Continue at top-level of nested iteration    for [i=a:b] for [j=c:*]
This is a second partial fix for Bug #2589
The first partial fix ensured that an inner unbounded iteration would
terminate when the source runs out of data.
This fix introduces a flag that indicates that has happened,
and allows an outer iteration to increment, reset the flag, and continue.

However this would mean that if two unbounded iterations are nested
the outer one would never terminate.  So
(1) We warn if multiple nested iterations are detected by "plot", and
(2) We try to catch an actual top-level overrun and stop early,
    which is probably not what the user expected.expected.

I am not entirely convinced that (2) is foolproof but at least there
will be a warning.