New plot style "with sectors"
This work introduces a new plotting style 'sectors' which plots
an annular sector segment for each input data point.
This is only relevant to 2D data plots.

The 'using' specifier provides either 4 or 6 basic columns of input data.
The shape of each sector is determined by 4 parameters, i.e. the coordinates of a
corner, the sector angle, and the radial width of annulus.

Columns 1 and 2 hold the coordinates of a corner on a conceptual polar grid.
The plot style itself can be used with either cartesian or polar coordinates;
in either case columns 1 and 2 are interpreted as azimuth and radius.
Columns 3 and 4 determine the angular and radial extent of the sector.
If no additional coordinates for the center of the annulus are given in
columns 5 and 6 for each point, it defaults to (0,0).
A per-sector color may be provided in the last column of the using specifier.

    using corner_azimuth:corner_radius:sector_angle:annular_width
    using corner_azimuth:corner_radius:sector_angle:annular_width:color
    using corner_azimuth:corner_radius:sector_angle:annular_width:center_x:center_y
    using corner_azimuth:corner_radius:sector_angle:annular_width:center_x:center_y:color