implement missing command "unset theta"
This was documented to return the theta convention to
it's default state
     set theta right ccw
but the command itself did not exist.  Now it does.

Instead theta was reset by "unset polar".
That would become a problem if the internal routine polar_to_xy()
is used in cartesian context.  It isn't at present but Hiroki Motoyoshi's
work on "with sector" changes that.