post.trm - modify outdated kludge for buggy ghostscript
Since 2003 (commits 647573ef 3601de122b) post.trm has had a
work-around for a bug in ghostscript that limited the number of
successive relative moves or lines.  Furthermore it added a "stroke"
command before continuing.  Because this causes a spurious line if it
occurs inside a newpath/closepath, the limit was set at 100 (later 250)
segments on the assumption that no closed path longer than that
would be generated.

That assumption is no longer true.  Now we do not apply this kludge
if a path is active.  Probably the entire mechanism is no longer
needed, but this change is more conservative than removing it altogether.