Bugfix: empty last field of a csv file incorrectly handled
If the final character on a line of a csv file is a field separator,
that field is marked as "missing".  However some string handling routines
did not check for the "missing" flag and as a consequence re-cycled an
previous string value, possibly leading to use-after-free errors.

Now we both clear df_column[].position, which pointed to the previous
string, and in general (df_parse_string_field()) substitute a new blank
string in place of a missing string.