Handle nonlinear axes during contour generation
(backport from 6.1)

The contour code dates back to before we switched to storing the true
coordinate values of log-scale data.  It "normalizes" the contour
segments by scaling them relative to |xmax-xmin| and |ymax-ymin|.
But this makes no sense any more for log-scale data.

For the default "set cntrparam linear", this makes a difference only
in deciding whether two points are so close together that there is no
need to draw a line between them.  This logic was in fuzzy_equal(),
which we now rewrite to handle nonlinear axes.  The bug reported was
that all contours in region of the low end of the log axis were not drawn.

Bug #2593