xerrorbars xyerrorbars: visually separate the point from the error bars
yerrorbars was already doing this.
Now all errorbar styles are handled the same.
A background-filled circle with size taken from "set pointintervalbox"
is positioned between the error bar and the point in front of it.
The size of the circle, i.e. the size of the visual gap in the bar,
is controlled by "set pointintervalbox".

Bug #2598  (not so much a bug as a feature request)