3D boxes finally get individual control over the fill border
In version 5 all 3D boxes shared the border color specified by
"set pm3d border <color>".  Now that addition of field
quadrangle.from_plot makes it easier to trace back to the actual
plot specification we can revise the code to retrieve per-plot
fillstyle details including border color.

Note:

Since version 5 did not support a distinction between
"linecolor" and "fillcolor" in the splot command, these were
effectively synonyms in many cases.  Now there is a distinction,
so v5 scripts that used "linecolor" to set the color of filled
areas may need to replace that keyword with "fillcolor".
I.e.
	splot DATA with boxes lc [rgb] variable
may need to become
	splot DATA with boxes fc [rgb] variable

TODO:

This leaves the border of 3D polygons controlled by the pm3d
setting rather than the fill style, which is an unforunate point
of difference between boxes and polygons.

Note that individual polygon objects are not associated with an
"splot" command so per-plot properties make no sense.
They do have a fillstyle in the object definition, but this is
not accessible to the pm3d code.

Possibly the quadrangle.fillstyle field could be replaced by a
pointer to a full struct fill_style_type?