re-purpose the fillstyle field in struct quadrangle to hold a zclip index
The conceptual basis for new plot style CONTOURFILL is that a single
plot structure describing a pm3d surface can be processed multiple times,
each time selecting a discrete band of z values to assign a single color.
To support this we need to store an index in the generated quadrangles
that indicates which pass through the surface it came from.

The plot header points to an array
	struct zslice zclip[]
in which each entry holds zclip limits zmin, zmax and a fill color
to use for the portion of the surface between those limits.
The array pointer remains constant as the surface is processed
multiple times, but obviously the index into the array must change.
The problem comes if/when "set pm3d depthorder" collects all the
quadrangles from all the iterations.  They all point back to the
same header, which is fine for the array itself, but the plot header
at that point holds only the final index value used.  So rather than
looking in the plot header for the appropriate index we must
store it in the quadrangle itself.

pm3d quadrangles that did not come from a surface, e.g. polygon
objects, still use this field to hold the fillstyle.
This change potentially those plot styles if if I got anything wrong,
so it needs a lot of testing.