New plot style splot with contourfill (working implementation)
Many thanks to Hiroki Motoyoshi for the observation that smooth
clipping on z in 3D is a workable approach to generating filled
contour plots in 2D.

This is an initial implementation of the plot style.
It supports commands

    set contourfill auto N
	split zrange evenly into N slices
    set contourfill ztics <level>
	slice boundary at each z-axis tick with matching tic level
    set contourfill cbtics <level>
	slice boundary at each cb-axis tick with matching tic level

    splot SURFACE with contourfill

In all cases the slice is colored by the palette color mapping for
the mean z of the slice.

A demo is included, mostly for debugging.
No documentation yet.

TODO:

- Some way to set colors other than using the palette.
  Maybe a separate set of sequence slice colors?
      set contourfill colors (color1, color2, color3, ...)
  Maybe a linestyle sequence, like "set cntrparam firstlinetype N"?

- take color from a colormap rather than from the current palette
	maybe:      set contourfill palette MyColors
	or maybe:   splot SURFACE with contourfill palette MyColors

- Not sure if anything special is needed to handle log-scale z or cb.

- documentation