new configure option --with-mergesort
Some libc versions, including macOS, provide a qsort that is unstable;
i.e. it does not guarantee to retain the previous ordering of elements
whose sort values are equal.  This can affect the output of hidden3d
surfaces and pm3d depthorder surfaces.  It potentially affects also
the zsort filter and convex/concave hull generation.

This option attempts to find mergesort in a standard library.
If found, it is used in place of qsort.

Not recommended unless your system has an unstable qsort and the
resulting inconsistent appearance of superimposed identical pm3d
or hidden3d surfaces bothers you.

Bug #2604 #2446