2nd try at a configuration option to request a stable sort for pm3d
This change subsumes the check for mergesort into a wider set of
alternatives explored by
	./configure --enable-stable-sort

The qsort implementation in glibc is stable; i.e. elements that test
as equal are guaranteed to retain their origin order.
The relevance to gnuplot is that if either "pm3d depthorder" or
"set hidden3d" is used to render two surfaces with identical [x,y,z],
the first will be consistently occluded by the second.

However the qsort provided by BSD and macOS is not stable.
Those systems may provide mergesort as a stable alternative.

Other non-glibc platforms including mingw and Cygwin do not provide a stable sort.
On these platforms, identical surfaces sorted for either pm3d or hidden3d
are rendered inconsistently.  Any given quadrangle or line segment may
be rendered with the properties of either of the requested surfaces.

Gnuplot will now try to pick an appropriate stable sort if possible
and otherwise add code so that pm3d depth-sorting is stable.
There is not yet a fallback for hidden3d processing if the platform
does not offer a stable sort.

./configure --with-stable-sort
     1) Use mergesort if the system provides it, otherwise qsort.
     2) If neither mergesort nor glibc's qsort is available
        add a sequence field to pm3d quadrangles to be used as a
	secondary sort key.

Bug #2604
See earlier commits b992a505 103cc301 9282ff70