fallback stable sort for hidden3d "with lines"
The previous commit c5530094 added a fallback implementation of a
stable sort algorithm for pm3d depth-sorting.  This commit adds a
similar fallback for hidden3d surfaces.

For plots generated by "splot with lines" the p_number field of
plot->lp_properties is now overloaded to hold a secondary sort key
for vertices derived from that plot.
This field is normally used only by style "with linespoints" and
is ignored by hidden3d processing in that case (not a change).

Bug #2604