docs:  Revise prolog to TeX documentation for use with texlive2023
texlive2023 seems much more sensitive to the ordering of package loading.
In order to get everything working for pdflatex as it had been for
texlive2022 and earlier I had to rearrange the sections for
utf8 encoding, character substitutions, hyperlinks, and indexing.

I also collected into one section all the options specific to pdflatex
and made a parallel section for options specific to lualatex/xelatex.

Switch between pdflatex and lualatex by adding comments to one
and removing comments from the other