bugfix: combination of pm3d options "at b" or "at t" with "depthorder"
Since the introduction of smooth z-clipping in version 5.4, pm3d quadrangles
drawn on the base plane always got clipped out of existence during depthorder
processing because the flag intended to protect them is always zero in this
case.

Now we instead use (quadrangle.corner[0].z == base_z) as a test for a base
quadrangle rather then trying to check the flag.  So the base plane is
drawn, but it remains a limitation that the component quadrangles cannot be
smoothly clipped against the original z limits.

This commit also fixes an oversight that if pm3d interpolation was active,
pm3d depth sorting placed components of both the top and bottom planes on
the surface instead of at base_z or ceiling_z.