Refactor base64 functions in write_png_image to take a callback
Instead of having a FILE* in the base64state and writing to it using
putc, there is now a void* and a pointer to a putc-like callback
function.

The existing write_png_base64_image still works the same as it always
has, but it's now a thin wrapper around the new write_png_base64_cb.