Do not warn if XDG directories cannot be opened because $HOME is not defined
If enviromental variable HOME is not set, which might be the case when
running a script anonymously for example from a cgi script, the program warns
     warning: HOME not set - cannot expand tilde
when looking for XDG configuration directories.  The warning is pointless,
and it is potentially harmful to continue a perhaps create a file somewhere
else instead.  Now we simply fail the XDG lookup silently.

Bug #2623