introduce NULL mousing routines for non-mousing terminals
Mostly this just guarantees that the program will not segfault if
it tries to call term->set_cursor() etc.

The interesting part is that it introduces a generic routine
term_waitforinput() that allows non-mousing terminals to filter input
from stdin through checks for key bindings including arrow keys.
At this point it is only active during "pause mouse" but the scope
can be widened later.

Example of use:
    set term sixelgd truecolor animate
    splot x*y with pm3d
    pause mouse
          arrow keys can now rotate the plot
	  'l' toggles log scale; 'g' toggles grid; etc
	  <cr> to return to normal command parsing.

It works nicely with sixelgd.
It sort of works with domterm but is too slow to be useable.

TODO:
 - written for use with libreadline;
   modify to support libedit and builtin readline
 - shift/control/alt modifiers not handled
 - in principle it could handle mouse events also but I have no idea
   how to catch them from inside an arbitrary terminal window.