Restore possibly swallowed input character when returning from "pause mouse"
There is a long history on this one.
See discussion on previous tracker issues
https://sourceforge.net/p/gnuplot/bugs/2541/
https://sourceforge.net/p/gnuplot/bugs/2292/
https://sourceforge.net/p/gnuplot/bugs/2279/
https://sourceforge.net/p/gnuplot/bugs/2134/
https://sourceforge.net/p/gnuplot/bugs/1559/

We are fighting two contrary problems

(1) When multiplexing input from the keyboard and mouse events,
sometimes the mouse event processing eats one character from stdin that
it doesn't know what to do with.  So in several places the code tries to
stuff this character back into the stdin buffer so it will be there later
for the normal command interpreter parsing to see.
In other places in expectation that the next character may or may not
be swallowed the code pushes a newline character back onto stdin, following
the logic that whether the newline is lost or not either way the first
character of the next command is left intact.

(2) However inserting a '\n' can cause premature return from a pause.

There may well be a solution to both problems simultaneously but so far
no one has suggested it.

Previous commits
   	 ccdb1936 67e3f16a9c
applied the stuff-back (1) only when "pause mouse" was not active.
The present commit does the same thing inside the "pause mouse" code.

Only lightly tested.  It is possible that this will re-break whatever
cases the previous commits were trying to avoid by testing for "pause mouse".