plugin demo:   plugin header has gotten out of sync with gp_types.h
The DATA_TYPES and structure definition in gnuplot_plugin.h must
match those in gp_types.h.  It says that right in the source, but
they no longer match.

This is bad.  There was a mismatch all through version 5.4.
Brought back into sync now, but there really should be a version
flag of some sort that the plugin loading code can test to confirm
that the plugin is compatible with the gnuplot version loading it.