Install man pages (en and ja) at canonical locations
Use man page names that are compatible with automake(1).
Otherwise they are not installed at all.
The man page system requires strict naming conventions and
file tree structure, which automake(1) adheres to.

This patch corrects the man page names, file tree, and makes the
complementary adjustments to the build system.

There is no need to distribute the same man page with multiple
encodings. The recommended best practice is to use UTF-8.

Patch tested for 'make', 'make install', and 'make dist-lzip' on
x86_64 Linux.

Resolves Bug #2230