revise code that draws 3D key box entries
Toggling 3D plots on/off by clicking the corresponding key entry
caused bad colors for both the key text and key samples of other entries.
The x11 and wxt terminals are affected because toggling is implemented by
skipping or not skipping the relevant graphics object.  This means that
the sequence of graphics operations changes if any plots are toggled off.
Qt is not affected because toggling is instead implemented by simply
setting the "hidden" attribute for those objects rather than skipping them.

This change re-orders the 3D code so that for any given plot or contour
the key text and key sample are drawn together, i.e. the graph itself
does not intervene.  Thus skipping the operations that make up the graph
does not change the sequence of operations in the key.

Bug #2636