post: handle RGBA images (only current use is to render a pixmap)
Since PostScript does not support transparency, gnuplot uses the
"with rgbalpha pixels" pixel-by-pixel emulation to handle the alpha
channel in images.  However pixmaps may have an alpha channel and
they bypass the general image processing by calling term->image()
directly.  This meant that the postscript terminal driver was getting
image data in mode IC_RGBA, which it did not support. Now it throws
away the alpha channel and treats the image as if it used mode IC_RGB.

Bug #2644